recipes/by-use-case/trading-copilot/beginner/requirements.txt
langgraph>=0.2.0
openai>=1.40.0
requests>=2.32.0
pyyaml>=6.0
yfinance>=0.2.40
pandas>=2.0.0
ta>=0.11.0
feedparser>=6.0.0
pytest>=8.0.0
